vrclient: Remove unnecessary return value cast.
CW-Bug-Id: #22729

+++ vrclient_x64/cppIVRIOBuffer_IVRIOBuffer_001.cpp
@@ -5,7 +5,7 @@ NTSTATUS IVRIOBuffer_IVRIOBuffer_001_Open( void *args )
 {
     struct IVRIOBuffer_IVRIOBuffer_001_Open_params *params = (struct IVRIOBuffer_IVRIOBuffer_001_Open_params *)args;
     struct u_IVRIOBuffer_IVRIOBuffer_001 *iface = (struct u_IVRIOBuffer_IVRIOBuffer_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->Open( params->pchPath, params->mode, params->unElementSize, params->unElements, params->pulBuffer );
+    params->_ret = iface->Open( params->pchPath, params->mode, params->unElementSize, params->unElements, params->pulBuffer );
     return 0;
 }
 
@@ -13,7 +13,7 @@ NTSTATUS IVRIOBuffer_IVRIOBuffer_001_Close( void *args )
 {
     struct IVRIOBuffer_IVRIOBuffer_001_Close_params *params = (struct IVRIOBuffer_IVRIOBuffer_001_Close_params *)args;
     struct u_IVRIOBuffer_IVRIOBuffer_001 *iface = (struct u_IVRIOBuffer_IVRIOBuffer_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->Close( params->ulBuffer );
+    params->_ret = iface->Close( params->ulBuffer );
     return 0;
 }
 
@@ -21,7 +21,7 @@ NTSTATUS IVRIOBuffer_IVRIOBuffer_001_Read( void *args )
 {
     struct IVRIOBuffer_IVRIOBuffer_001_Read_params *params = (struct IVRIOBuffer_IVRIOBuffer_001_Read_params *)args;
     struct u_IVRIOBuffer_IVRIOBuffer_001 *iface = (struct u_IVRIOBuffer_IVRIOBuffer_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->Read( params->ulBuffer, params->pDst, params->unBytes, params->punRead );
+    params->_ret = iface->Read( params->ulBuffer, params->pDst, params->unBytes, params->punRead );
     return 0;
 }
 
@@ -29,7 +29,7 @@ NTSTATUS IVRIOBuffer_IVRIOBuffer_001_Write( void *args )
 {
     struct IVRIOBuffer_IVRIOBuffer_001_Write_params *params = (struct IVRIOBuffer_IVRIOBuffer_001_Write_params *)args;
     struct u_IVRIOBuffer_IVRIOBuffer_001 *iface = (struct u_IVRIOBuffer_IVRIOBuffer_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->Write( params->ulBuffer, params->pSrc, params->unBytes );
+    params->_ret = iface->Write( params->ulBuffer, params->pSrc, params->unBytes );
     return 0;
 }
 
@@ -37,7 +37,7 @@ NTSTATUS IVRIOBuffer_IVRIOBuffer_001_PropertyContainer( void *args )
 {
     struct IVRIOBuffer_IVRIOBuffer_001_PropertyContainer_params *params = (struct IVRIOBuffer_IVRIOBuffer_001_PropertyContainer_params *)args;
     struct u_IVRIOBuffer_IVRIOBuffer_001 *iface = (struct u_IVRIOBuffer_IVRIOBuffer_001 *)params->linux_side;
-    params->_ret = (uint64_t)iface->PropertyContainer( params->ulBuffer );
+    params->_ret = iface->PropertyContainer( params->ulBuffer );
     return 0;
 }
