+++ vrclient_x64/cppIVRMailbox_IVRMailbox_001.cpp
@@ -5,7 +5,7 @@ NTSTATUS IVRMailbox_IVRMailbox_001_undoc1( void *args )
 {
     struct IVRMailbox_IVRMailbox_001_undoc1_params *params = (struct IVRMailbox_IVRMailbox_001_undoc1_params *)args;
     struct u_IVRMailbox_IVRMailbox_001 *iface = (struct u_IVRMailbox_IVRMailbox_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->undoc1( params->a, params->b );
+    params->_ret = iface->undoc1( params->a, params->b );
     return 0;
 }
 
@@ -13,7 +13,7 @@ NTSTATUS IVRMailbox_IVRMailbox_001_undoc2( void *args )
 {
     struct IVRMailbox_IVRMailbox_001_undoc2_params *params = (struct IVRMailbox_IVRMailbox_001_undoc2_params *)args;
     struct u_IVRMailbox_IVRMailbox_001 *iface = (struct u_IVRMailbox_IVRMailbox_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->undoc2( params->a );
+    params->_ret = iface->undoc2( params->a );
     return 0;
 }
 
@@ -21,7 +21,7 @@ NTSTATUS IVRMailbox_IVRMailbox_001_undoc4( void *args )
 {
     struct IVRMailbox_IVRMailbox_001_undoc4_params *params = (struct IVRMailbox_IVRMailbox_001_undoc4_params *)args;
     struct u_IVRMailbox_IVRMailbox_001 *iface = (struct u_IVRMailbox_IVRMailbox_001 *)params->linux_side;
-    params->_ret = (uint32_t)iface->undoc4( params->a, params->b, params->c, params->d );
+    params->_ret = iface->undoc4( params->a, params->b, params->c, params->d );
     return 0;
 }
